complete challenge 2
- compute the estimate for hospital beds
- add result to estimate and severeImpact return object

[Delivers 4#issue-597905962]
